refactor(resume-parser): move functionality to shared
