+++ src/shared/resume-parser/extract-resume-from-sections/lib/bullet-points.ts
