+++ src/shared/resume-parser/extract-resume-from-sections/lib/common-features.ts
